Added prod and dev environment support and reorganized files in preparation for MongoDB migration
