+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 package-lock.json
-.env
+.env.*
 userData.json
 dist
 .idea
