+++ src/games/zenless_zone_zero/checkin_zenless.ts
@@ -1,13 +1,14 @@
-import {Profile} from "../bot";
+import {Profile} from "../../models";
 
-export async function hkstrCheckin(profile: Profile): Promise<string> {
-    
-    const url = 'https://sg-public-api.hoyolab.com/event/luna/os/sign?act_id=e202303301540311';
+export async function zzzCheckin(profile: Profile): Promise<string> {
+
+    const url = 'https://sg-act-nap-api.hoyolab.com/event/luna/zzz/os/sign?lang=en-us&act_id=e202406031448091';
     const username = profile.nickname;
     const cookies = 'ltoken_v2='+profile.pasted_cookie.ltoken_v2+';'+'ltuid_v2='+profile.pasted_cookie.ltuid_v2+';';
 
     if (!url) {
-        return `Check-in skipped for ${username}: Honkai Starrail check-in is disabled.`;
+        return `Check-in skipped for ${username}: Zenless Zone Zero check-in is disabled.`;
+
     }
 
     const header = {
@@ -21,12 +22,12 @@ export async function hkstrCheckin(profile: Profile): Promise<string> {
         'Referer': 'https://act.hoyolab.com/',
         'Origin': 'https://act.hoyolab.com',
     };
-    
+
     const options: RequestInit = {
         method: 'POST',
         headers: header,
     };
-  
+
     // Check-in the profile
     try {
         const hoyolabResponse: Response = await fetch(url, options);
